demo/requirements.txt
streamlit
pandas
numpy
matplotlib
spacy
pillow
requests